Restructure the controller
