+++ packages/backend/src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { APIController } from './api.controller';
-import { UsersModule } from './users/users.module';
-import { HouseModule } from './house/house.module';
+import { UsersModule } from './controllers/users/users.module';
+import { HouseModule } from './controllers/house/house.module';
 import { MongooseModule } from '@nestjs/mongoose';
 import { PassportModule } from '@nestjs/passport';
