+++ packages/backend/src/controllers/house/dto/create-house.dto.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { HouseModel } from '../../../db/house/house.schema';
+
+export class CreateHouseDto extends HouseModel {
+  @ApiProperty()
+  readonly name: string;
+
+  @ApiProperty()
+  readonly email: string;
+
+  @ApiProperty()
+  readonly address: string;
+}
