+++ packages/backend/src/controllers/house/dto/join-house.dto.ts
@@ -1,0 +1,6 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class JoinHouseDto {
+  @ApiProperty()
+  readonly houseCode: string;
+}
