+++ packages/backend/src/controllers/house/house.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { HouseController } from './house.controller';
-import { DbModule } from '../db/db.module';
+import { DbModule } from '../../db/db.module';
 import { HouseUtil } from './house.util';
 import { FirebaseAuthStrategy } from 'src/guards/firebase.auth';
