+++ packages/backend/src/controllers/users/dto/create-user.dto.ts
@@ -1,0 +1,9 @@
+import { UserModel } from '../../../db/user/user.schema';
+import { Schema as MongooseSchema } from 'mongoose';
+import { ApiProperty } from '@nestjs/swagger';
+export class CreateUserDto extends UserModel {
+  house: MongooseSchema.Types.ObjectId;
+
+  @ApiProperty()
+  firebaseId: string;
+}
