+++ packages/backend/src/controllers/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
-import { DbModule } from '../db/db.module';
+import { DbModule } from '../../db/db.module';
 
 @Module({
   controllers: [UsersController],
