--- packages/backend/src/house/dto/create-house.dto.ts
@@ -1,3 +1,0 @@
-import { HouseModel } from '../../db/house/house.schema';
-
-export class CreateHouseDto extends HouseModel {}
