--- packages/backend/src/house/dto/join-house.dto.ts
@@ -1,3 +1,0 @@
-export class JoinHouseDto {
-  readonly houseCode: string;
-}
