--- packages/backend/src/users/dto/create-user.dto.ts
@@ -1,3 +1,0 @@
-import { UserModel } from '../../db/user/user.schema';
-
-export class CreateUserDto extends UserModel {}
